refactor: separate in progress page
