+++ client/src/lib/welcomeScreen.svelte
@@ -1,13 +1,38 @@
 <script lang="ts">
-  import { createNewGame } from './gameState.svelte';
+  import { gameState } from './gameState.svelte';
+
+  const { startNewGame, enterGame } = $props();
 </script>
 
-<div class="mb-4 flex items-center gap-2 bg-slate-200 p-4 dark:bg-slate-700">
+<h1 class="text-3xl font-bold">Welcome to Cardnames!</h1>
+{#if gameState.game === null || gameState.game.status === 'finished'}
+  <div class="mb-4 flex items-center gap-2 bg-slate-200 p-4 dark:bg-slate-700">
+    <p class="flex-grow">
+      There is no game currently in progress. Would you like to create a new game?
+    </p>
+    <button
+      class="rounded-md bg-sky-200 px-4 py-2 hover:bg-sky-300 active:bg-sky-100 dark:bg-sky-700 dark:hover:bg-sky-800 dark:active:bg-sky-600"
+      onclick={startNewGame}
+    >
+      Create game
+    </button>
+  </div>
+{:else if gameState.game.status === 'ready'}
   <p class="flex-grow">
-    There is no game currently in progress. Would you like to create a new game?
+    There is a game waiting to be started. Click the button below to join the game.
   </p>
   <button
     class="rounded-md bg-sky-200 px-4 py-2 hover:bg-sky-300 active:bg-sky-100 dark:bg-sky-700 dark:hover:bg-sky-800 dark:active:bg-sky-600"
-    onclick={createNewGame}>Create game</button
+    onclick={enterGame}
   >
-</div>
+    Join game
+  </button>
+{:else}
+  <p class="flex-grow">There is a game in progress. Click the button below to join the game.</p>
+  <button
+    class="rounded-md bg-sky-200 px-4 py-2 hover:bg-sky-300 active:bg-sky-100 dark:bg-sky-700 dark:hover:bg-sky-800 dark:active:bg-sky-600"
+    onclick={enterGame}
+  >
+    Join game
+  </button>
+{/if}
